\name{Italian_W}
\alias{Italian_W}
\title{
Italian provinces spatial weights matrix example
}
\description{
This is an example dataset that contains the 107 Italian provinces contiguity matrix (year 2008).
}
\usage{data(Italian_W)}
\format{
  A data frame with 107 x 107 row-standardized distances between observations (Italian provinces).
  }
\references{
http://www.istat.it/it/archivio/104317#confini.
}
\examples{
data(Italian_W)
}
\keyword{datasets}
