\name{ssfa-package}
\alias{ssfa-package}
\title{
Spatial Stochastic Frontier models
}
\description{
The package implements the Spatial Stochastic Frontier model for cross-sectional data introduced by Fusco and Vidoli (2013).
The method controls spatial heterogeneity in SFA models by splitting the inefficiency term into three parts: the first one related to spatial peculiarities of the territory in which each single unit operates, the second one related to the specific production features and the third one representing the error term.
}
\details{
\tabular{ll}{
Package: \tab ssfa\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2022-01-05\cr
License: \tab GPL-3 \cr
}
}
\author{
Elisa Fusco, Francesco Vidoli

Maintainer: Elisa Fusco <fusco_elisa@libero.it>
}
\references{
Fusco, E. and Vidoli, F. (2013). \emph{Spatial stochastic frontier models: controlling spatial global and local heterogeneity}, International Review of Applied Economics, 27(5) 679-694.
}
\keyword{ package }

